backend/requirements.txt
# Wav2Lip Course Generation Backend
# Python 3.8+ required

# Core dependencies
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0

# PyTorch - use compatible versions for Wav2Lip
# Install PyTorch separately based on your CUDA version:
# CPU: pip install torch torchvision torchaudio
# CUDA 11.8: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# CUDA 12.1: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0

# Wav2Lip core dependencies
numpy>=1.19.0
opencv-python>=4.5.0
opencv-contrib-python>=4.5.0
librosa>=0.8.0
scipy>=1.7.0
numba>=0.53.0

# Face detection for Wav2Lip
face-alignment>=1.3.5

# TTS (Text-to-Speech)

elevenlabs>=0.2.0

# Video processing
ffmpeg-python>=0.2.0
moviepy>=1.0.3
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Whisper for transcription
openai-whisper>=20231117

# First Order Motion Model (FOMM) dependencies
pyyaml>=6.0
scikit-image>=0.19.0
pandas>=1.3.0

# API and async processing
redis>=5.0.0
celery>=5.3.0
requests>=2.31.0

# Utilities
tqdm>=4.60.0
pillow>=9.0.0

# Authentication
sqlalchemy>=2.0.0
pyjwt>=2.8.0
bcrypt>=4.1.0
